+++ apps/backend/lambdas/users/package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "@aws-sdk/client-cognito-identity-provider": "^3.995.0",
     "@branch/lambda-auth": "file:../../../../shared/lambda-auth",
+    "@branch/lambda-http": "file:../../../../shared/lambda-http",
     "aws-jwt-verify": "^5.1.1",
     "kysely": "^0.28.8",
     "pg": "^8.16.3"
